cmake/install: add THRUST_INSTALL option
This determines whether an install target will be provided. If Thrust is
built as top-level project, the install target will be provided by default,
as before.

If Thrust is built as a sub-project, the install target will, by default,
not be provided, again maintaining existing behavior.

So what's new here is that via this option a downstream project can enable
install of thrust if it is used via `add_subdirectory`.